Implemented __getstate__ and  __setstate__ on LabelledData
Custom options are now restored to Store.custom_options when an object
that has been customized is unpickled.
